feat: Introduce appointment and message domain layers
Introduces the domain layer for managing appointments and messages, including repositories, data models, and use cases.

Key changes:
- Add `AppointmentRepository` for all appointment-related data operations.
- Add `MessageRepository` for handling messages linked to appointments.
- Define data models for `Appointment`, `Message`, `AppointmentWithDetails`, and `DashboardStats`.
- Implement use cases for creating, fetching, updating, and removing appointments.
- Expand `VisitorRepository` with methods to fetch visitors.
- Update `Business` and `Visitor` data models with new fields and adjust ID types to Long.
- Refactor `BusinessRepository` to remove visitor logic and use Long for IDs.

+++ composeApp/src/commonMain/kotlin/xyz/sattar/javid/proqueue/domain/usecase/DeleteBusinessUseCase.kt
@@ -3,6 +3,6 @@ package xyz.sattar.javid.proqueue.domain.usecase
 import xyz.sattar.javid.proqueue.domain.BusinessRepository
 
 class DeleteBusinessUseCase(private val repository: BusinessRepository) {
-    suspend operator fun invoke(businessId: Int) =
+    suspend operator fun invoke(businessId: Long) =
         repository.deleteBusiness(businessId)
 }
